add extra mastra wrapper with defaults
