+++ packages/core/src/action/index.ts
@@ -2,7 +2,7 @@ import type { z } from 'zod';
 
 import type { Agent } from '../agent';
 import type { Logger } from '../logger';
-import type { Mastra } from '../mastra';
+import type { Mastra } from '../mastra/mastra';
 import type { MastraMemory } from '../memory';
 import type { MastraStorage } from '../storage';
 import type { Telemetry } from '../telemetry';
