+++ packages/core/src/mastra/mastra.ts
@@ -1,0 +1,421 @@
+import type { Agent } from '../agent';
+import type { MastraDeployer } from '../deployer';
+import { noopLogger } from '../logger';
+import type { Logger } from '../logger';
+import type { MastraMemory } from '../memory/memory';
+import type { MastraStorage } from '../storage';
+import { InstrumentClass, Telemetry } from '../telemetry';
+import type { OtelConfig } from '../telemetry';
+import type { MastraTTS } from '../tts';
+import type { MastraVector } from '../vector';
+import type { Workflow } from '../workflows';
+
+export interface Config<
+  TAgents extends Record<string, Agent<any>> = Record<string, Agent<any>>,
+  TWorkflows extends Record<string, Workflow> = Record<string, Workflow>,
+  TVectors extends Record<string, MastraVector> = Record<string, MastraVector>,
+  TTTS extends Record<string, MastraTTS> = Record<string, MastraTTS>,
+  TLogger extends Logger = Logger,
+> {
+  agents?: TAgents;
+  storage: MastraStorage;
+  vectors?: TVectors;
+  logger: TLogger | false;
+  workflows?: TWorkflows;
+  tts?: TTTS;
+  telemetry?: OtelConfig;
+  deployer?: MastraDeployer;
+
+  /**
+   * Server middleware functions to be applied to API routes
+   * Each middleware can specify a path pattern (defaults to '/api/*')
+   */
+  serverMiddleware?: Array<{
+    handler: (c: any, next: () => Promise<void>) => Promise<Response | void>;
+    path?: string;
+  }>;
+
+  /**
+   * @deprecated add memory to your Agent directly instead
+   */
+  memory?: MastraMemory;
+}
+
+@InstrumentClass({
+  prefix: 'mastra',
+  excludeMethods: ['getLogger', 'getTelemetry'],
+})
+export class Mastra<
+  TAgents extends Record<string, Agent<any>> = Record<string, Agent<any>>,
+  TWorkflows extends Record<string, Workflow> = Record<string, Workflow>,
+  TVectors extends Record<string, MastraVector> = Record<string, MastraVector>,
+  TTTS extends Record<string, MastraTTS> = Record<string, MastraTTS>,
+> {
+  #vectors?: TVectors;
+  #agents: TAgents;
+  #logger: Logger;
+  #workflows: TWorkflows;
+  #tts?: TTTS;
+  #deployer?: MastraDeployer;
+  #serverMiddleware: Array<{
+    handler: (c: any, next: () => Promise<void>) => Promise<Response | void>;
+    path: string;
+  }> = [];
+  #telemetry?: Telemetry;
+  #storage?: MastraStorage;
+  #memory?: MastraMemory;
+
+  /**
+   * @deprecated use getTelemetry() instead
+   */
+  get telemetry() {
+    return this.#telemetry;
+  }
+
+  /**
+   * @deprecated use getStorage() instead
+   */
+  get storage() {
+    return this.#storage;
+  }
+
+  /**
+   * @deprecated use getMemory() instead
+   */
+  get memory() {
+    return this.#memory;
+  }
+
+  constructor(config: Config<TAgents, TWorkflows, TVectors, TTTS>) {
+    // Store server middleware with default path
+    if (config.serverMiddleware) {
+      this.#serverMiddleware = config.serverMiddleware.map(m => ({
+        handler: m.handler,
+        path: m.path || '/api/*',
+      }));
+    }
+
+    let logger: Logger = config.logger ? config.logger : (noopLogger as unknown as Logger);
+    this.#logger = logger;
+
+    /*
+    Telemetry
+    */
+    this.#telemetry = Telemetry.init(config?.telemetry);
+
+    /*
+      Storage
+    */
+    let storage = config?.storage;
+    this.#storage = storage;
+
+    if (this.#telemetry) {
+      this.#storage = this.#telemetry.traceClass(storage, {
+        excludeMethods: ['__setTelemetry', '__getTelemetry'],
+      });
+      this.#storage.__setTelemetry(this.#telemetry);
+    }
+
+    /*
+    Vectors
+    */
+    if (config.vectors) {
+      let vectors: Record<string, MastraVector> = {};
+      Object.entries(config.vectors).forEach(([key, vector]) => {
+        if (this.#telemetry) {
+          vectors[key] = this.#telemetry.traceClass(vector, {
+            excludeMethods: ['__setTelemetry', '__getTelemetry'],
+          });
+          vectors[key].__setTelemetry(this.#telemetry);
+        } else {
+          vectors[key] = vector;
+        }
+      });
+
+      this.#vectors = vectors as TVectors;
+    }
+
+    if (config.memory) {
+      this.#memory = config.memory;
+      if (this.#telemetry) {
+        this.#memory = this.#telemetry.traceClass(config.memory, {
+          excludeMethods: ['__setTelemetry', '__getTelemetry'],
+        });
+        this.#memory.__setTelemetry(this.#telemetry);
+      }
+    }
+
+    if (config && `memory` in config) {
+      this.#logger.warn(`
+  Memory should be added to Agents, not to Mastra.
+
+Instead of:
+  new Mastra({ memory: new Memory() })
+
+do:
+  new Agent({ memory: new Memory() })
+
+This is a warning for now, but will throw an error in the future
+`);
+    }
+
+    if (config?.tts) {
+      let tts: Record<string, MastraTTS> = {};
+      Object.entries(config.tts).forEach(([key, ttsCl]) => {
+        if (this.#telemetry) {
+          tts[key] = this.#telemetry.traceClass(ttsCl, {
+            excludeMethods: ['__setTelemetry', '__getTelemetry'],
+          });
+          tts[key].__setTelemetry(this.#telemetry);
+        } else {
+          tts[key] = ttsCl;
+        }
+      });
+
+      this.#tts = tts as TTTS;
+    }
+
+    /*
+    Agents
+    */
+    const agents: Record<string, Agent> = {};
+    if (config?.agents) {
+      Object.entries(config.agents).forEach(([key, agent]) => {
+        if (agents[key]) {
+          throw new Error(`Agent with name ID:${key} already exists`);
+        }
+
+        agent.__registerPrimitives({
+          logger: this.getLogger(),
+          telemetry: this.#telemetry,
+          storage: this.storage,
+          memory: this.memory,
+          agents: agents,
+          tts: this.#tts,
+          vectors: this.#vectors,
+        });
+
+        agent.__registerMastra(this);
+
+        agents[key] = agent;
+      });
+    }
+
+    this.#agents = agents as TAgents;
+
+    /*
+    Workflows
+    */
+    this.#workflows = {} as TWorkflows;
+
+    if (config?.workflows) {
+      Object.entries(config.workflows).forEach(([key, workflow]) => {
+        workflow.__registerMastra(this);
+        workflow.__registerPrimitives({
+          logger: this.getLogger(),
+          telemetry: this.#telemetry,
+          storage: this.storage,
+          memory: this.memory,
+          agents: agents,
+          tts: this.#tts,
+          vectors: this.#vectors,
+        });
+        // @ts-ignore
+        this.#workflows[key] = workflow;
+      });
+    }
+
+    this.setLogger({ logger });
+  }
+
+  public getAgent<TAgentName extends keyof TAgents>(name: TAgentName): TAgents[TAgentName] {
+    const agent = this.#agents?.[name];
+    if (!agent) {
+      throw new Error(`Agent with name ${String(name)} not found`);
+    }
+    return this.#agents[name];
+  }
+
+  public getAgents() {
+    return this.#agents;
+  }
+
+  public getVector<TVectorName extends keyof TVectors>(name: TVectorName): TVectors[TVectorName] {
+    const vector = this.#vectors?.[name];
+    if (!vector) {
+      throw new Error(`Vector with name ${String(name)} not found`);
+    }
+    return vector;
+  }
+
+  public getVectors() {
+    return this.#vectors;
+  }
+
+  public getDeployer() {
+    return this.#deployer;
+  }
+
+  public getWorkflow<TWorkflowId extends keyof TWorkflows>(
+    id: TWorkflowId,
+    { serialized }: { serialized?: boolean } = {},
+  ): TWorkflows[TWorkflowId] {
+    const workflow = this.#workflows?.[id];
+    if (!workflow) {
+      throw new Error(`Workflow with ID ${String(id)} not found`);
+    }
+
+    if (serialized) {
+      return { name: workflow.name } as TWorkflows[TWorkflowId];
+    }
+
+    return workflow;
+  }
+
+  public getWorkflows(props: { serialized?: boolean } = {}): Record<string, Workflow> {
+    if (props.serialized) {
+      return Object.entries(this.#workflows).reduce((acc, [k, v]) => {
+        return {
+          ...acc,
+          [k]: { name: v.name },
+        };
+      }, {});
+    }
+    return this.#workflows;
+  }
+
+  public setStorage(storage: MastraStorage) {
+    this.#storage = storage;
+  }
+
+  public setLogger({ logger }: { logger: Logger }) {
+    this.#logger = logger;
+
+    if (this.#agents) {
+      Object.keys(this.#agents).forEach(key => {
+        this.#agents?.[key]?.__setLogger(this.#logger);
+      });
+    }
+
+    if (this.#memory) {
+      this.#memory.__setLogger(this.#logger);
+    }
+
+    if (this.#deployer) {
+      this.#deployer.__setLogger(this.#logger);
+    }
+
+    if (this.#tts) {
+      Object.keys(this.#tts).forEach(key => {
+        this.#tts?.[key]?.__setLogger(this.#logger);
+      });
+    }
+
+    if (this.#storage) {
+      this.#storage.__setLogger(this.#logger);
+    }
+
+    if (this.#vectors) {
+      Object.keys(this.#vectors).forEach(key => {
+        this.#vectors?.[key]?.__setLogger(this.#logger);
+      });
+    }
+  }
+
+  public setTelemetry(telemetry: OtelConfig) {
+    this.#telemetry = Telemetry.init(telemetry);
+
+    if (this.#agents) {
+      Object.keys(this.#agents).forEach(key => {
+        if (this.#telemetry) {
+          this.#agents?.[key]?.__setTelemetry(this.#telemetry);
+        }
+      });
+    }
+
+    if (this.#memory) {
+      this.#memory = this.#telemetry.traceClass(this.#memory, {
+        excludeMethods: ['__setTelemetry', '__getTelemetry'],
+      });
+      this.#memory.__setTelemetry(this.#telemetry);
+    }
+
+    if (this.#deployer) {
+      this.#deployer = this.#telemetry.traceClass(this.#deployer, {
+        excludeMethods: ['__setTelemetry', '__getTelemetry'],
+      });
+      this.#deployer.__setTelemetry(this.#telemetry);
+    }
+
+    if (this.#tts) {
+      let tts = {} as Record<string, MastraTTS>;
+      Object.entries(this.#tts).forEach(([key, ttsCl]) => {
+        if (this.#telemetry) {
+          tts[key] = this.#telemetry.traceClass(ttsCl, {
+            excludeMethods: ['__setTelemetry', '__getTelemetry'],
+          });
+          tts[key].__setTelemetry(this.#telemetry);
+        }
+      });
+      this.#tts = tts as TTTS;
+    }
+
+    if (this.#storage) {
+      this.#storage = this.#telemetry.traceClass(this.#storage, {
+        excludeMethods: ['__setTelemetry', '__getTelemetry'],
+      });
+      this.#storage.__setTelemetry(this.#telemetry);
+    }
+
+    if (this.#vectors) {
+      let vectors = {} as Record<string, MastraVector>;
+      Object.entries(this.#vectors).forEach(([key, vector]) => {
+        if (this.#telemetry) {
+          vectors[key] = this.#telemetry.traceClass(vector, {
+            excludeMethods: ['__setTelemetry', '__getTelemetry'],
+          });
+          vectors[key].__setTelemetry(this.#telemetry);
+        }
+      });
+      this.#vectors = vectors as TVectors;
+    }
+  }
+
+  public getTTS() {
+    return this.#tts;
+  }
+
+  public getLogger() {
+    return this.#logger;
+  }
+
+  public getTelemetry() {
+    return this.#telemetry;
+  }
+
+  public getMemory() {
+    return this.#memory;
+  }
+
+  public getStorage() {
+    return this.#storage;
+  }
+
+  public getServerMiddleware() {
+    return this.#serverMiddleware;
+  }
+
+  public async getLogsByRunId({ runId, transportId }: { runId: string; transportId: string }) {
+    if (!transportId) {
+      throw new Error('Transport ID is required');
+    }
+    return await this.#logger.getLogsByRunId({ runId, transportId });
+  }
+
+  public async getLogs(transportId: string) {
+    if (!transportId) {
+      throw new Error('Transport ID is required');
+    }
+    return await this.#logger.getLogs(transportId);
+  }
+}
