+++ packages/core/src/workflows/machine.ts
@@ -9,7 +9,7 @@ import type { z } from 'zod';
 import type { MastraUnion } from '../action';
 import type { Logger } from '../logger';
 
-import type { Mastra } from '../mastra';
+import type { Mastra } from '../mastra/mastra';
 import { createMastraProxy } from '../utils';
 import type { Step } from './step';
 import type {
